+++ geotrek/settings/env_dev.py
@@ -2,6 +2,15 @@
 # Django Development
 # ..........................
 
+import sys
+import os
+
+# Makemessages uses Locale_paths's settings to get the list of translations to do.
+# We change 'var/extra_locale' to 'geotrek/locale'
+
+if 'makemessages' in sys.argv:
+    LOCALE_PATHS = (os.path.join(PROJECT_DIR, 'locale'),)
+
 DEBUG = True
 
 EMAIL_BACKEND = 'django.core.mail.backends.console.EmailBackend'