+++ docs/changelog.rst
@@ -3344,7 +3344,8 @@ In order to enable those features under construction, add ``experimental = True`
 :notes:
 
     Give related permissions to the managers group in order to allow edition
-    (``add_flatpage``, ``change_flatpage``, ``delete_flatpage``, ``add_touristiccontent`` ...).
+    (``add_flatpage``, ``change_flatpage``, ``delete_flatpage``,
+     ``add_touristiccontent`` ...).
 
 
 0.27.2 (2010-10-14)
@@ -3528,7 +3529,6 @@ Since the map export have changed, empty the cache :
 * Rework display of lists in detail pages, better factorization
 * Removed links in logbook list for certain models
 * Display messages in login page too (useful for redirections)
-
 Support edition of several fields on the same map, via django-leaflet new feature (fixes #53)